network: remove keepalived
Initially, I installed keepalived to have an automatic failover in case
of master failure. It doesn't seem possible to configure keepalived to
failover after ~x minutes when the master is dead* (see below).

Since servers will host sensitive data on a PostgreSQL database, the
automatic failover might be risky and I prefer to have monitoring and a
manual failover.

* the initial goal was to have automatic failover: on reboot, I wanted
the master to keep being the master and have an outage of several
minutes ; but in case of real outage, I wanted the slave to failover. To
achieve this, it is possible to set the variable "advert_int" of
keepalived. With advert_int=180, slave becomes master if master is down
more than ~180 seconds, but in case of master reboot, it doesn't become
master again before ~180 seconds after keepalived is up. In short,
configuring keepalived to achieve a failover after x minutes or hours
doesn't seem possible.